requirement.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
click==8.1.8
Django==5.1.7
django-cors-headers==4.7.0
django-extensions==4.1
django-ninja==1.4.1
djangorestframework==3.15.2
dlib==19.24.6
drf-yasg==1.21.10
face-recognition==1.3.0
face_recognition_models @ git+https://github.com/ageitgey/face_recognition_models@e67de717267507d1e9246de95692eb8be736ab61
fastapi==0.115.12
idna==3.10
inflection==0.5.1
numpy==2.2.4
packaging==24.2
pillow==11.1.0
pydantic==2.11.3
pydantic_core==2.33.1
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
setuptools==78.1.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.12.2
uritemplate==4.1.1
